+++ modules/mux/mpeg/tsutil.c
@@ -27,7 +27,7 @@
 #include "tsutil.h"
 
 void PEStoTS( void *p_opaque, PEStoTSCallback pf_callback, block_t *p_pes,
-              int i_pid, bool *pb_discontinuity, int *pi_continuity_counter )
+              uint16_t i_pid, bool *pb_discontinuity, uint8_t *pi_continuity_counter )
 {
     /* get PES total size */
     uint8_t *p_data = p_pes->p_buffer;
