+++ modules/mux/mpeg/tsutil.h
@@ -23,6 +23,6 @@
 typedef void(*PEStoTSCallback)(void *, block_t *);
 
 void PEStoTS( void *p_opaque, PEStoTSCallback pf_callback, block_t *p_pes,
-              int i_pid, bool *pb_discontinuity, int *pi_continuity_counter );
+              uint16_t i_pid, bool *pb_discontinuity, uint8_t *pi_continuity_counter );
 
 #endif
